+++ apps/rush-lib/src/cli/actions/PublishAction.ts
@@ -19,12 +19,12 @@ import { Npm } from '../../utilities/Npm';
 import { RushCommandLineParser } from '../RushCommandLineParser';
 import { PublishUtilities } from '../../logic/PublishUtilities';
 import { ChangelogGenerator } from '../../logic/ChangelogGenerator';
-import { GitPolicy } from '../../logic/GitPolicy';
 import { PrereleaseToken } from '../../logic/PrereleaseToken';
 import { ChangeManager } from '../../logic/ChangeManager';
 import { BaseRushAction } from './BaseRushAction';
-import { Git } from '../../logic/Git';
+import { PublishGit } from '../../logic/PublishGit';
 import { VersionControl } from '../../utilities/VersionControl';
+import { PolicyValidator } from '../../logic/policy/PolicyValidator';
 
 export class PublishAction extends BaseRushAction {
   private _addCommitDetails: CommandLineFlagParameter;
@@ -173,29 +173,28 @@ export class PublishAction extends BaseRushAction {
    * Executes the publish action, which will read change request files, apply changes to package.jsons,
    */
   protected run(): Promise<void> {
-    if (!GitPolicy.getUserEmail(this.rushConfiguration, true)) {
-      process.exit(1);
-      return Promise.resolve();
-    }
-    const allPackages: Map<string, RushConfigurationProject> = this.rushConfiguration.projectsByName;
+    return Promise.resolve().then(() => {
+      PolicyValidator.validatePolicy(this.rushConfiguration, false);
 
-    if (this._regenerateChangelogs.value) {
-      console.log('Regenerating changelogs');
-      ChangelogGenerator.regenerateChangelogs(allPackages, this.rushConfiguration);
-      return Promise.resolve();
-    }
+      const allPackages: Map<string, RushConfigurationProject> = this.rushConfiguration.projectsByName;
 
-    this._validate();
+      if (this._regenerateChangelogs.value) {
+        console.log('Regenerating changelogs');
+        ChangelogGenerator.regenerateChangelogs(allPackages, this.rushConfiguration);
+        return Promise.resolve();
+      }
 
-    if (this._includeAll.value) {
-      this._publishAll(allPackages);
-    } else {
-      this._prereleaseToken = new PrereleaseToken(this._prereleaseName.value, this._suffix.value);
-      this._publishChanges(allPackages);
-    }
+      this._validate();
 
-    console.log(EOL + colors.green('Rush publish finished successfully.'));
-    return Promise.resolve();
+      if (this._includeAll.value) {
+        this._publishAll(allPackages);
+      } else {
+        this._prereleaseToken = new PrereleaseToken(this._prereleaseName.value, this._suffix.value);
+        this._publishChanges(allPackages);
+      }
+
+      console.log(EOL + colors.green('Rush publish finished successfully.'));
+    });
   }
 
   /**
@@ -224,7 +223,7 @@ export class PublishAction extends BaseRushAction {
 
     if (changeManager.hasChanges()) {
       const orderedChanges: IChangeInfo[] = changeManager.changes;
-      const git: Git = new Git(this._targetBranch.value);
+      const git: PublishGit = new PublishGit(this._targetBranch.value);
       const tempBranch: string = 'publish-' + new Date().getTime();
 
       // Make changes in temp branch.
@@ -279,7 +278,7 @@ export class PublishAction extends BaseRushAction {
     console.log(`Rush publish starts with includeAll and version policy ${this._versionPolicy.value}`);
 
     let updated: boolean = false;
-    const git: Git = new Git(this._targetBranch.value);
+    const git: PublishGit = new PublishGit(this._targetBranch.value);
 
     allPackages.forEach((packageConfig, packageName) => {
       if (packageConfig.shouldPublish &&
@@ -303,7 +302,7 @@ export class PublishAction extends BaseRushAction {
     }
   }
 
-  private _gitAddTags(git: Git, orderedChanges: IChangeInfo[]): void {
+  private _gitAddTags(git: PublishGit, orderedChanges: IChangeInfo[]): void {
     for (const change of orderedChanges) {
       if (
         change.changeType &&
@@ -388,7 +387,11 @@ export class PublishAction extends BaseRushAction {
       const destFolder: string = this._releaseFolder.value ?
        this._releaseFolder.value : path.join(this.rushConfiguration.commonTempFolder, 'artifacts', 'packages');
 
-      FileSystem.move(tarballPath, path.join(destFolder, tarballName), { overwrite: true });
+      FileSystem.move({
+        sourcePath: tarballPath,
+        destinationPath: path.join(destFolder, tarballName),
+        overwrite: true
+      });
     }
   }
 
@@ -417,7 +420,10 @@ export class PublishAction extends BaseRushAction {
           const toApiFolderPath: string = path.join(project.projectFolder, toApiFolder);
           if (FileSystem.exists(fromApiFolderPath) && FileSystem.exists(toApiFolderPath)) {
             FileSystem.readFolder(fromApiFolderPath).forEach(fileName => {
-              FileSystem.copyFile(path.join(fromApiFolderPath, fileName), path.join(toApiFolderPath, fileName));
+              FileSystem.copyFile({
+                sourcePath: path.join(fromApiFolderPath, fileName),
+                destinationPath: path.join(toApiFolderPath, fileName)
+              });
               console.log(`Copied file ${fileName} from ${fromApiFolderPath} to ${toApiFolderPath}`);
             });
           }