+++ build-tests/api-extractor-test-03/src/index.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+/// <reference path="../typings/tsd.d.ts" />
+
 /**
  * api-extractor-test-03
  *