+++ common/scripts/install-run.js
@@ -394,5 +394,4 @@ function run() {
     });
 }
 run();
-
 //# sourceMappingURL=install-run.js.map