+++ libraries/node-core-library/src/FileSystem.ts
@@ -5,113 +5,17 @@ import * as pathUtilities from 'path';
 import * as fs from 'fs';
 import * as fsx from 'fs-extra';
 
-import { Text, NewlineKind } from './Text';
+import { Text, NewlineKind, Encoding } from './Text';
+import { PosixModeBits } from './PosixModeBits';
 
 // The PosixModeBits are intended to be used with bitwise operations.
 // tslint:disable:no-bitwise
 
 /**
- * The allowed types of encodings, as supported by Node.js
- * @public
- */
-export const enum Encoding {
-  Utf8 = 'utf8'
-}
-
-/**
- * An integer value used to specify file permissions for POSIX-like operating systems.
- *
- * @remarks
- *
- * This bitfield corresponds to the "mode_t" structure described in this document:
- * http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_stat.h.html
- *
- * It is used with NodeJS APIs such as fs.Stat.mode and fs.chmodSync().  These values
- * represent a set of permissions and can be combined using bitwise arithmetic.
- *
- * POSIX is a registered trademark of the Institute of Electrical and Electronic Engineers, Inc.
- *
- * @public
- */
-export const enum PosixModeBits {
-  // The bits
-
-  /**
-   * Indicates that the item's owner can read the item.
-   */
-  UserRead = 1 << 8,
-
-  /**
-   * Indicates that the item's owner can modify the item.
-   */
-  UserWrite = 1 << 7,
-
-  /**
-   * Indicates that the item's owner can execute the item (if it is a file)
-   * or search the item (if it is a directory).
-   */
-  UserExecute = 1 << 6,
-
-  /**
-   * Indicates that users belonging to the item's group can read the item.
-   */
-  GroupRead = 1 << 5,
-
-  /**
-   * Indicates that users belonging to the item's group can modify the item.
-   */
-  GroupWrite = 1 << 4,
-
-  /**
-   * Indicates that users belonging to the item's group can execute the item (if it is a file)
-   * or search the item (if it is a directory).
-   */
-  GroupExecute = 1 << 3,
-
-  /**
-   * Indicates that other users (besides the item's owner user or group) can read the item.
-   */
-  OthersRead = 1 << 2,
-
-  /**
-   * Indicates that other users (besides the item's owner user or group) can modify the item.
-   */
-  OthersWrite = 1 << 1,
-
-  /**
-   * Indicates that other users (besides the item's owner user or group) can execute the item (if it is a file)
-   * or search the item (if it is a directory).
-   */
-  OthersExecute = 1 << 0,
-
-  // Helpful aliases
-
-  /**
-   * A zero value where no permissions bits are set.
-   */
-  None = 0,
-
-  /**
-   * An alias combining OthersRead, GroupRead, and UserRead permission bits.
-   */
-  AllRead = OthersRead | GroupRead | UserRead,
-
-  /**
-   * An alias combining OthersWrite, GroupWrite, and UserWrite permission bits.
-   */
-  AllWrite = OthersWrite | GroupWrite | UserWrite,
-
-  /**
-   * An alias combining OthersExecute, GroupExecute, and UserExecute permission bits.
-   */
-  AllExecute = OthersExecute | GroupExecute | UserExecute
-}
-
-/**
  * The options for FileSystem.readFolder()
  * @public
  */
-export interface IReadFolderOptions {
+export interface IFileSystemReadFolderOptions {
   /**
    * If true, returns the absolute paths of the files in the folder.
    * Defaults to `false`.
@@ -123,7 +27,7 @@ export interface IReadFolderOptions {
  * The options for FileSystem.writeFile()
  * @public
  */
-export interface IWriteFileOptions {
+export interface IFileSystemWriteFileOptions {
   /**
    * If true, will ensure the folder is created before writing the file.
    * Defaults to `false`.
@@ -147,7 +51,7 @@ export interface IWriteFileOptions {
  * The options for FileSystem.readFile()
  * @public
  */
-export interface IReadFileOptions {
+export interface IFileSystemReadFileOptions {
   /**
    * If specified, will change the encoding of the file that will be written.
    * Defaults to `"utf8"`.
@@ -167,6 +71,18 @@ export interface IReadFileOptions {
  */
 export interface IFileSystemMoveOptions {
   /**
+   * The path of the existing object to be moved.
+   * The path may be absolute or relative.
+   */
+  sourcePath: string;
+
+  /**
+   * The new path for the object.
+   * The path may be absolute or relative.
+   */
+  destinationPath: string;
+
+  /**
    * If true, will overwrite the file if it already exists. Defaults to true.
    */
   overwrite?: boolean;
@@ -179,10 +95,28 @@ export interface IFileSystemMoveOptions {
 }
 
 /**
+ * The options for FileSystem.copyFile()
+ * @public
+ */
+export interface IFileSystemCopyFileOptions {
+  /**
+   * The path of the existing object to be copied.
+   * The path may be absolute or relative.
+   */
+  sourcePath: string;
+
+  /**
+   * The path that the object will be copied to.
+   * The path may be absolute or relative.
+   */
+  destinationPath: string;
+}
+
+/**
  * The options for FileSystem.deleteFile()
  * @public
 */
-export interface IDeleteFileOptions {
+export interface IFileSystemDeleteFileOptions {
   /**
    * If true, will throw an exception if the file did not exist before `deleteFile()` was called.
    * Defaults to `false`.
@@ -195,7 +129,7 @@ export interface IDeleteFileOptions {
  * Both times must be specified.
  * @public
  */
-export interface IUpdateTimeParameters {
+export interface IFileSystemUpdateTimeParameters {
   /**
    * The POSIX epoch time or Date when this was last accessed.
    */
@@ -205,6 +139,24 @@ export interface IUpdateTimeParameters {
    * The POSIX epoch time or Date when this was last modified
    */
   modifiedTime: number | Date;
+}
+
+/**
+ * The options for `FileSystem.createSymbolicLinkJunction()`, `createSymbolicLinkFile()`,
+ * `createSymbolicLinkFolder()`,  and `createHardLink()`.
+ *
+ * @public
+ */
+export interface IFileSystemCreateLinkOptions {
+  /**
+   * The existing path that the symbolic link will point to.
+   */
+  linkTargetPath: string;
+
+  /**
+   * The new path for the new symlink link to be created.
+   */
+  newLinkPath: string;
 }
 
 /**
@@ -219,7 +171,7 @@ export interface IUpdateTimeParameters {
  * performance, versus improving it.
  *
  * Note that in the documentation, we refer to "filesystem objects", this can be a
- * file, folder, synbolic link, hard link, directory junction, etc.
+ * file, folder, symbolic link, hard link, directory junction, etc.
  *
  * @public
  */
@@ -261,7 +213,7 @@ export class FileSystem {
    * @param path - The path of the file that should be modified.
    * @param times - The times that the object should be updated to reflect.
    */
-  public static updateTimes(path: string, times: IUpdateTimeParameters): void {
+  public static updateTimes(path: string, times: IFileSystemUpdateTimeParameters): void {
     // tslint:disable-next-line:no-any
     fsx.utimes(path, times.accessedTime as any, times.modifiedTime as any);
   }
@@ -313,11 +265,8 @@ export class FileSystem {
   /**
    * Moves a file. The folder must exist, unless the `ensureFolderExists` option is provided.
    * Behind the scenes it uses `fs-extra.moveSync()`
-   * @param sourcePath - The absolute or relative path to the source file.
-   * @param targetPath - The absolute or relative path where the file should be moved to.
-   * @param options - Optional settings that can change the behavior. Type: `IFileSystemMoveOptions`
    */
-  public static move(sourcePath: string, targetPath: string, options?: IFileSystemMoveOptions): void {
+  public static move(options: IFileSystemMoveOptions): void {
     options = {
       overwrite: true,
       ensureFolderExists: false,
@@ -325,10 +274,10 @@ export class FileSystem {
     };
 
     if (options.ensureFolderExists) {
-      FileSystem.ensureFolder(pathUtilities.basename(sourcePath));
+      FileSystem.ensureFolder(pathUtilities.basename(options.sourcePath));
     }
 
-    fsx.moveSync(sourcePath, targetPath, { overwrite: options.overwrite });
+    fsx.moveSync(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
   }
 
   // ===============
@@ -352,7 +301,7 @@ export class FileSystem {
    * @param folderPath - The absolute or relative path to the folder which should be read.
    * @param options - Optional settings that can change the behavior. Type: `IReadFolderOptions`
    */
-  public static readFolder(folderPath: string, options?: IReadFolderOptions): Array<string> {
+  public static readFolder(folderPath: string, options?: IFileSystemReadFolderOptions): Array<string> {
     options = {
       absolutePaths: false,
       ...options
@@ -407,7 +356,7 @@ export class FileSystem {
    * @param contents - The text that should be written to the file.
    * @param options - Optional settings that can change the behavior. Type: `IWriteFileOptions`
    */
-  public static writeFile(filePath: string, contents: string | Buffer, options?: IWriteFileOptions): void {
+  public static writeFile(filePath: string, contents: string | Buffer, options?: IFileSystemWriteFileOptions): void {
     options = {
       ensureFolderExists: false,
       convertLineEndings: undefined,
@@ -431,7 +380,7 @@ export class FileSystem {
    * @param filePath - The relative or absolute path to the file whose contents should be read.
    * @param options - Optional settings that can change the behavior. Type: `IReadFileOptions`
    */
-  public static readFile(filePath: string, options?: IReadFileOptions): string {
+  public static readFile(filePath: string, options?: IFileSystemReadFileOptions): string {
     options = {
       encoding: Encoding.Utf8,
       convertLineEndings: undefined,
@@ -455,11 +404,9 @@ export class FileSystem {
    * Copies a file from one location to another.
    * By default, destinationPath is overwritten if it already exists.
    * Behind the scenes it uses `fs.copyFileSync()`.
-   * @param sourcePath - The absolute or relative path to the source file to be copied.
-   * @param destinationPath - The absolute or relative path to the new copy that will be created.
    */
-  public static copyFile(sourcePath: string, destinationPath: string): void {
-    fsx.copySync(sourcePath, destinationPath);
+  public static copyFile(options: IFileSystemCopyFileOptions): void {
+    fsx.copySync(options.sourcePath, options.destinationPath);
   }
 
   /**
@@ -468,7 +415,7 @@ export class FileSystem {
    * @param filePath - The absolute or relative path to the file that should be deleted.
    * @param options - Optional settings that can change the behavior. Type: `IDeleteFileOptions`
    */
-  public static deleteFile(filePath: string, options?: IDeleteFileOptions): void {
+  public static deleteFile(filePath: string, options?: IFileSystemDeleteFileOptions): void {
     options = {
       throwIfNotExists: false,
       ...options
@@ -501,42 +448,34 @@ export class FileSystem {
   /**
    * Creates a Windows "directory junction". Behaves like `createSymbolicLinkToFile()` on other platforms.
    * Behind the scenes it uses `fs.symlinkSync()`.
-   * @param linkSource - The absolute or relative path to the destination where the link should be created.
-   * @param linkTarget - The absolute or relative path to the target of the link.
    */
-  public static createSymbolicLinkJunction(linkTarget: string, linkSource: string): void {
+  public static createSymbolicLinkJunction(options: IFileSystemCreateLinkOptions): void {
     // For directories, we use a Windows "junction".  On POSIX operating systems, this produces a regular symlink.
-    fsx.symlinkSync(linkTarget, linkSource, 'junction');
+    fsx.symlinkSync(options.linkTargetPath, options.newLinkPath, 'junction');
   }
 
   /**
    * Creates a symbolic link to a file (on Windows this requires elevated permissionsBits).
    * Behind the scenes it uses `fs.symlinkSync()`.
-   * @param linkSource - The absolute or relative path to the destination where the link should be created.
-   * @param linkTarget - The absolute or relative path to the target of the link.
    */
-  public static createSymbolicLinkFile(linkTarget: string, linkSource: string): void {
-    fsx.symlinkSync(linkSource, linkTarget, 'file');
+  public static createSymbolicLinkFile(options: IFileSystemCreateLinkOptions): void {
+    fsx.symlinkSync(options.linkTargetPath, options.newLinkPath, 'file');
   }
 
   /**
    * Creates a symbolic link to a folder (on Windows this requires elevated permissionsBits).
    * Behind the scenes it uses `fs.symlinkSync()`.
-   * @param linkSource - The absolute or relative path to the destination where the link should be created.
-   * @param linkTarget - The absolute or relative path to the target of the link.
    */
-  public static createSymbolicLinkFolder(linkTarget: string, linkSource: string): void {
-    fsx.symlinkSync(linkSource, linkTarget, 'dir');
+  public static createSymbolicLinkFolder(options: IFileSystemCreateLinkOptions): void {
+    fsx.symlinkSync(options.linkTargetPath, options.newLinkPath, 'dir');
   }
 
   /**
    * Creates a hard link.
    * Behind the scenes it uses `fs.linkSync()`.
-   * @param linkSource - The absolute or relative path to the destination where the link should be created.
-   * @param linkTarget - The absolute or relative path to the target of the link.
    */
-  public static createHardLink(linkTarget: string, linkSource: string): void {
-    fsx.linkSync(linkSource, linkTarget);
+  public static createHardLink(options: IFileSystemCreateLinkOptions): void {
+    fsx.linkSync(options.linkTargetPath, options.newLinkPath);
   }
 
   /**