+++ apps/api-extractor/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @microsoft/api-extractor
 
-This log was last generated on Wed, 22 Aug 2018 20:58:58 GMT and should not be manually modified.
+This log was last generated on Wed, 29 Aug 2018 06:36:50 GMT and should not be manually modified.
+
+## 5.10.8
+Wed, 29 Aug 2018 06:36:50 GMT
+
+*Version update only*
+
+## 5.10.7
+Thu, 23 Aug 2018 18:18:53 GMT
+
+### Patches
+
+- Republish all packages in web-build-tools to resolve GitHub issue #782
 
 ## 5.10.6
 Wed, 22 Aug 2018 20:58:58 GMT