+++ core-build/gulp-core-build-mocha/CHANGELOG.md
@@ -1,6 +1,33 @@
 # Change Log - @microsoft/gulp-core-build-mocha
 
-This log was last generated on Wed, 22 Aug 2018 20:58:58 GMT and should not be manually modified.
+This log was last generated on Thu, 30 Aug 2018 19:23:16 GMT and should not be manually modified.
+
+## 3.5.7
+Thu, 30 Aug 2018 19:23:16 GMT
+
+*Version update only*
+
+## 3.5.6
+Thu, 30 Aug 2018 18:45:12 GMT
+
+*Version update only*
+
+## 3.5.5
+Wed, 29 Aug 2018 21:43:23 GMT
+
+*Version update only*
+
+## 3.5.4
+Wed, 29 Aug 2018 06:36:50 GMT
+
+*Version update only*
+
+## 3.5.3
+Thu, 23 Aug 2018 18:18:53 GMT
+
+### Patches
+
+- Republish all packages in web-build-tools to resolve GitHub issue #782
 
 ## 3.5.2
 Wed, 22 Aug 2018 20:58:58 GMT