+++ core-build/gulp-core-build-typescript/CHANGELOG.md
@@ -1,6 +1,42 @@
 # Change Log - @microsoft/gulp-core-build-typescript
 
-This log was last generated on Wed, 22 Aug 2018 20:58:58 GMT and should not be manually modified.
+This log was last generated on Thu, 30 Aug 2018 22:47:34 GMT and should not be manually modified.
+
+## 4.11.12
+Thu, 30 Aug 2018 22:47:34 GMT
+
+### Patches
+
+- Include defaultTslint in npm package.
+
+## 4.11.11
+Thu, 30 Aug 2018 19:23:16 GMT
+
+*Version update only*
+
+## 4.11.10
+Thu, 30 Aug 2018 18:45:12 GMT
+
+*Version update only*
+
+## 4.11.9
+Wed, 29 Aug 2018 21:43:23 GMT
+
+*Version update only*
+
+## 4.11.8
+Wed, 29 Aug 2018 06:36:50 GMT
+
+### Patches
+
+- Fixing minor path issue.
+
+## 4.11.7
+Thu, 23 Aug 2018 18:18:53 GMT
+
+### Patches
+
+- Republish all packages in web-build-tools to resolve GitHub issue #782
 
 ## 4.11.6
 Wed, 22 Aug 2018 20:58:58 GMT