+++ core-build/node-library-build/CHANGELOG.md
@@ -1,6 +1,40 @@
 # Change Log - @microsoft/node-library-build
 
-This log was last generated on Wed, 22 Aug 2018 20:58:58 GMT and should not be manually modified.
+This log was last generated on Thu, 30 Aug 2018 22:47:34 GMT and should not be manually modified.
+
+## 5.0.4
+Thu, 30 Aug 2018 22:47:34 GMT
+
+*Version update only*
+
+## 5.0.3
+Thu, 30 Aug 2018 19:23:16 GMT
+
+*Version update only*
+
+## 5.0.2
+Thu, 30 Aug 2018 18:45:12 GMT
+
+*Version update only*
+
+## 5.0.1
+Wed, 29 Aug 2018 21:43:23 GMT
+
+*Version update only*
+
+## 5.0.0
+Wed, 29 Aug 2018 06:36:50 GMT
+
+### Breaking changes
+
+- Updating the way typescript and tslint are invoked.
+
+## 4.4.11
+Thu, 23 Aug 2018 18:18:53 GMT
+
+### Patches
+
+- Republish all packages in web-build-tools to resolve GitHub issue #782
 
 ## 4.4.10
 Wed, 22 Aug 2018 20:58:58 GMT